feat(auth): Add KeyID
Signing processes can add the key ID into the client data to communicate,
which key was used for the signature. Verifying processes can use this
information to check if they have the right key. This allows us to rotate
the key on a regular basis and distinguish between "verification failed"
and "we have to fetch the new key".